Multi Language- Creating the Default Language Seeder
 Multi Language- Creating the Default Language Seeder

+++ app/Http/Controllers/Admin/LanguageController.php
@@ -89,6 +89,9 @@ class LanguageController extends Controller
     {
         try {
             $language = Language::findOrFail($id);
+            if($language->lang === 'en'){
+                return response(['status' => 'error', 'message' => __('admin.Can\'t Delete This One!')]);
+            }
             $language->delete();
             return response(['status' => 'success', 'message' => __('admin.Deleted Successfully!')]);
         } catch (\Throwable $th) {
